+++ yuyo/backoff.py
@@ -58,6 +58,8 @@ class Backoff:
     this will either back off for the time passed to
     [yuyo.backoff.Backoff.set_next_backoff][] if applicable or a time calculated exponentially.
 
+    Each iteration yields the current retry count (starting at 0).
+
     Examples
     --------
     An example of using this class as an asynchronous iterator may look like
@@ -155,14 +157,17 @@ class Backoff:
     def __aiter__(self) -> Backoff:
         return self
 
-    async def __anext__(self) -> None:
+    async def __anext__(self) -> int:
         # We don't want to backoff on the first iteration.
         if not self._started:
             self._started = True
-            return
+            return 0
 
-        if await self.backoff():
+        result = await self.backoff()
+        if result is None:
             raise StopAsyncIteration
+
+        return result
 
     @property
     def is_depleted(self) -> bool:
@@ -173,21 +178,21 @@ class Backoff:
         """
         return self._max_retries is not None and self._max_retries == self._retries
 
-    async def backoff(self) -> bool:
+    async def backoff(self) -> typing.Optional[int]:
         """Sleep for the provided backoff or for the next exponent.
 
         This provides an alternative to iterating over this class.
 
         Returns
         -------
-        bool
+        int | None
             Whether this has reached the end of its iteration.
 
             If this returns [True][] then that call didn't sleep as this has
             been marked as finished or has reached the max retries.
         """
         if self._finished or self.is_depleted:
-            return True
+            return None
 
         self._started = True
         # We do this even if _next_backoff is set to make sure it's always incremented.
@@ -198,7 +203,7 @@ class Backoff:
 
         self._retries += 1
         await asyncio.sleep(backoff_)
-        return False
+        return self._retries
 
     def finish(self) -> None:
         """Mark the iterator as finished to break out of the current loop."""
@@ -267,7 +272,7 @@ class ErrorManager:
     def __init__(
         self,
         *rules: typing.Tuple[
-            typing.Iterable[typing.Type[BaseException]], typing.Callable[[typing.Any], typing.Optional[bool]]
+            typing.Iterable[typing.Type[Exception]], typing.Callable[[typing.Any], typing.Optional[bool]]
         ],
     ) -> None:
         """Initialise an error manager instance.
@@ -317,7 +322,7 @@ class ErrorManager:
 
     def with_rule(
         self,
-        exceptions: typing.Iterable[typing.Type[BaseException]],
+        exceptions: typing.Iterable[typing.Type[Exception]],
         result: typing.Callable[[typing.Any], typing.Optional[bool]],
     ) -> Self:
         """Add a rule to this exception context manager.